+++ gandiDns/CHANGELOG.md
@@ -1,5 +1,16 @@
-## 0.0.1
-Frist release version
+# Changelog
 
-## 0.0.2
-Updating documentation
+
+## [1.0.0] - 2022-06-03
+### Fixed
+- Fixed the limitation "when updating a record, the api, remove ALL entries for this record in the DNS zone file"
+### Changed
+- Fixed typo of the config parameter reccords -> records
+
+## [0.0.2] - 2021-01-07
+### Changed
+- Updating documentation
+
+## [0.0.1] - 2020-12-12
+### Added
+- First release version